requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
scikit-learn>=1.3.0
pillow>=10.0.0
tqdm>=4.65.0
flask>=3.0.0
flask-cors>=4.0.0
groq>=0.4.0
python-dotenv>=1.0.0
joblib>=1.3.0
jupyter>=1.0.0
ipykernel>=6.25.0
